Move delay out of simple directory
