--- src/lib/delay/index.ts
@@ -1,1 +1,0 @@
-export { default as Simple } from './simple/Simple.svelte'
